refactored api code to be generalisable
